fix journey tests by allowing cargo package to go online
`--offline` doesn't work anymore as the default registry is now sparse, which
has nothing cached that it didn't see before.

`--offline` probably was an optimization that we just can't have anymore.

The alternative would be to switch to git, but that proabbly would mean a whole
clone on CI as well, or failure as `--offline` would prevent that anyway.